teams prefabs, damage, getting damage

+++ Assets/Scripts/Interfaces/IDamagable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamagable
 {
-    void GetDamage(int damageValue);
+    void GetDamage(float damageValue);
 }
